+++ onmt/modules/Decoder.lua
@@ -333,19 +333,27 @@ function Decoder:backward(batch, outputs, criterion)
   for t = batch.targetLength, 1, -1 do
     -- Compute decoder output gradients.
     -- Note: This would typically be in the forward pass.
+    _G.profiler:start("generator.fwd")
     local pred = self.generator:forward(outputs[t])
+    _G.profiler:stop("generator.fwd")
     local output = batch:getTargetOutput(t)
 
+    _G.profiler:start("criterion.fwd")
     loss = loss + criterion:forward(pred, output)
+    _G.profiler:stop("criterion.fwd")
 
     -- Compute the criterion gradient.
+    _G.profiler:start("criterion.bwd")
     local genGradOut = criterion:backward(pred, output)
+    _G.profiler:stop("criterion.bwd")
     for j = 1, #genGradOut do
       genGradOut[j]:div(batch.totalSize)
     end
 
     -- Compute the final layer gradient.
+    _G.profiler:start("generator.bwd")
     local decGradOut = self.generator:backward(outputs[t], genGradOut)
+    _G.profiler:stop("generator.bwd")
     gradStatesInput[#gradStatesInput]:add(decGradOut)
 
     -- Compute the standarad backward.