+++ onmt/utils/init.lua
@@ -13,5 +13,6 @@ utils.Parallel = require('onmt.utils.Parallel')
 utils.Features = require('onmt.utils.Features')
 utils.Log = require('onmt.utils.Log')
 utils.Logger = require('onmt.utils.Logger')
+utils.Profiler = require('onmt.utils.Profiler')
 
 return utils